add aws-iot-greengrass-lite-webrtc-demo-image
This image contain amazon-kvs-webrtc-sdk and when started it automatically
connect to Amazon Kinesis Video Streams WebRTC using the kvsWebrtcClientMasterGstSample
that use the default camera and microphone to init a WebRTC connection as a Master.
It is based on aws-iot-greengrass-lite-demo-image and does have the same features,
such as A/B update, read only partition etc.